Edit and add volunteers implemented, avatar pictures recovered from gravatar
