+++ src/app/meetings/check-in/check-in.component.html
@@ -20,7 +20,7 @@
       <tr *ngFor="let volunteer of assistants">
         <td>
           <h4 class="ui image header">
-            <img src="./assets/img/no-photo.jpeg" class="ui mini rounded image">
+            <img [src]="volunteer.pictureURL" class="ui mini rounded image">
             <div class="content">
               {{volunteer.name}}
             </div>
