+++ src/app/meetings/meetings.component.ts
@@ -1,6 +1,5 @@
 import { Component } from '@angular/core';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
 
 import { MeetingsService } from './meetings.service';
 import { IMeeting } from './i-meeting';
@@ -14,15 +13,6 @@ export class MeetingsComponent {
   meetings$: Observable<Array<IMeeting>>;
 
   constructor(private meetingsService: MeetingsService) {
-    this.meetings$ = this.meetingsService.getMeetings().pipe(
-      map(actions =>
-        actions.map(a => {
-          const data = a.payload.doc.data() as IMeeting;
-          const date = data.date as any;
-          data.date = new Date(date.seconds * 1000);
-          return data;
-        })
-      )
-    );
+    this.meetings$ = this.meetingsService.getMeetings();
   }
 }
